+++ src/zope/interface/interface.py
@@ -352,9 +352,10 @@ class InterfaceClass(Element, InterfaceBase, Specification):
         # Make sure that all recorded attributes (and methods) are of type
         # `Attribute` and `Method`
         for name, attr in list(attrs.items()):
-            if name in ('__locals__', '__qualname__'):
+            if name in ('__locals__', '__qualname__', '__annotations__'):
                 # __locals__: Python 3 sometimes adds this.
                 # __qualname__: PEP 3155 (Python 3.3+)
+                # __annotations__: PEP 3107 (Python 3.0+)
                 del attrs[name]
                 continue
             if isinstance(attr, Attribute):